Makefile:  Fix more places where attempts to use Unix-style paths with native Windows toolchain.  tools/mksymtab.sh:  Add a check just to prevent 'cat' of an empty file list from failing.  Right not the root problem is that the are not executables being built or installed in the exe/ directory.  That needs to be done by 'make import' but is not happening.

+++ tools/mksymtab.sh
@@ -69,12 +69,14 @@ rm -f $outfile
 # list of sorted, unique undefined variable names.
 
 execlist=`find ${dir} -type f`
-for exec in ${execlist}; do
-  nm $exec | fgrep ' U ' | sed -e "s/^[ ]*//g" | cut -d' ' -f2  >>_tmplist
-done
+if [ ! -z "${execlist}"]; then
+  for exec in ${execlist}; do
+    nm $exec | fgrep ' U ' | sed -e "s/^[ ]*//g" | cut -d' ' -f2  >>_tmplist
+  done
 
-varlist=`cat _tmplist | sort - | uniq -`
-rm -f _tmplist
+  varlist=`cat _tmplist | sort - | uniq -`
+  rm -f _tmplist
+fi
 
 # Now output the symbol table as a structure in a C source file.  All
 # undefined symbols are declared as void* types.  If the toolchain does
